Nest+Nuxt Authorization

+++ server/src/job/job.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Param, Post, Res, UseInterceptors, UploadedFile, Body, ValidationPipe, Delete, ParseIntPipe, Patch, Query } from '@nestjs/common';
+import { Controller, Get, Param, Post, Res, UseInterceptors, UploadedFile, Body, ValidationPipe, Delete, ParseIntPipe, Patch, Query, UseGuards, UnauthorizedException } from '@nestjs/common';
 import { FileInterceptor } from '@nestjs/platform-express'
 import { diskStorage } from 'multer'
 import { JobService } from './job.service';
@@ -6,13 +6,20 @@ import { Job } from './job.entity';
 import { editFileName, imageFileFilter } from 'src/utils/image-uploading.utils';
 import { CreateJobDto } from './dto/create-job.dto';
 import { GetJobsPaginationDto } from './dto/get-jobs-pagination.dto'
+import { AuthGuard } from '@nestjs/passport';
+import { GetCompany } from 'src/company/get-company.decorator';
+import { Company } from 'src/company/company.entity';
 
 @Controller('job')
 export class JobController {
     constructor(private jobService: JobService) { }
 
     @Get('admin')
-    async getJobs(): Promise<Job[]> {
+    @UseGuards(AuthGuard('jwt'))
+    async getJobs(@GetCompany() company: Company): Promise<Job[]> {
+        if (!company.admin) {
+            throw new UnauthorizedException('Invalid credentials')
+        }
         return this.jobService.getJobs()
     }
 
@@ -29,6 +36,7 @@ export class JobController {
     }
 
     @Post()
+    @UseGuards(AuthGuard('jwt'))
     @UseInterceptors(FileInterceptor('image', {
         storage: diskStorage({
             destination: './files',
@@ -36,18 +44,37 @@ export class JobController {
         }),
         fileFilter: imageFileFilter
     }))
-    async createJob(@UploadedFile() file,
+    async createJob(
+        @GetCompany() company: Company,
+        @UploadedFile() file,
         @Body(ValidationPipe) createJobDto: CreateJobDto): Promise<void> {
+        if (!company.admin) {
+            throw new UnauthorizedException('Invalid credentials')
+        }
         return this.jobService.createJob(file, createJobDto)
     }
 
     @Patch(':id/approve')
-    async approveJob(@Param('id', ParseIntPipe) id: number, @Body() dataObj: { email: string }): Promise<void> {
+    @UseGuards(AuthGuard('jwt'))
+    async approveJob(
+        @GetCompany() company: Company,
+        @Param('id', ParseIntPipe) id: number,
+        @Body() dataObj: { email: string }): Promise<void> {
+        if (!company.admin) {
+            throw new UnauthorizedException('Invalid credentials')
+        }
         return this.jobService.approveJob(id, dataObj.email)
     }
 
     @Delete(':id')
-    async deleteJob(@Param('id', ParseIntPipe) id: number, @Body() dataObj: { email: string }): Promise<void> {
+    @UseGuards(AuthGuard('jwt'))
+    async deleteJob(
+        @GetCompany() company: Company,
+        @Param('id', ParseIntPipe) id: number,
+        @Body() dataObj: { email: string }): Promise<void> {
+        if (!company.admin) {
+            throw new UnauthorizedException('Invalid credentials')
+        }
         return this.jobService.deleteJob(id, dataObj.email)
     }
 }
